requirements.txt
PyYAML==6.0.3
requests==2.32.5
pandas~=2.3.3
numpy~=2.4.0
scikit-learn~=1.8.0